p7zip: Added package
Command line (de)compression tool for 7z archives.
  http://p7zip.sourceforge.net/

Only the "7zr" tool is built, which is sufficient to handle 7z
archives, but doesn't support all the other compression formats that
the full 7-Zip and p7zip packages do.

Optional optimizations such as CPU-specific compile flags and assembly
routines, as well as huge pages, are not used in this build.