sdl_mixer: Enable support for libmodplug
SDL_mixer will prefer libmodplug to mikmod if both are enabled.
Which is a good thing, since we experienced crashes with mikmod on
some songs on the GCW Zero.

The patch fixes the include path: since libmodplug 0.8.8.5, the
cflags reported by pkg-config no longer add the "libmodplug" dir,
so the #include directive must contain that path instead.